[REF] backend: serve static file with withenoise in local dev
there is 3 environments
- debug
- local
- prod

in debug, the static files are served by django
in local, they are not ; it is expected to have a reverse proxy (nginx)
in prod, they are served by s3

since we don't need (afaict right now) nginx on production since we have
an ALB before gunicorn, it seems to me configuring a second docker image
with nginx a bit overkill. We thus decide to use whitenoise only in
local mode, so that django serves the static files in a somewhat
performant way.